pyproject.toml
[project]
name = "src"
version = "0.0.1"
description = "UNICEF Future Texbooks project"
authors = [
    { name = "Ignacio Aristimuño", email = "nacho@marvik.ai" },
]
readme = "README.md"

requires-python = ">=3.10,<4.0"
dependencies = [
    "langgraph>=0.2.6",
    "langchain>=0.2.14",
    "langchain-experimental>=0.2.1",
    "langchain-core>=0.2.14",
    "langchain-groq>=0.2.14",
    "langchain-openai>=0.2.14",
    "langchain-community>=0.2.17",
    "python-dotenv>=1.0.1",
    "pandas>=2.1.1",
    "langgraph-cli[inmem] (>=0.1.71,<0.2.0)",
    "fastapi (>=0.115.10,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "retry>=0.9.2",
    "unidecode (>=1.3.8,<2.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.react_agent", "react_agent"]
[tool.setuptools.package-dir]
"langgraph.templates.react_agent" = "src/workflow"
"react_agent" = "src/workflow"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

